requirements.txt
# ── Core deep learning ────────────────────────────────────────────────────────
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# ── YOLO & supervision (ByteTrack) ────────────────────────────────────────────
ultralytics>=8.2.0          # YOLOv8m + YOLO training utilities
supervision>=0.21.0          # ByteTrack, annotation helpers

# ── Computer vision ───────────────────────────────────────────────────────────
opencv-python>=4.9.0
Pillow>=10.0.0
albumentations>=1.4.0        # Training augmentations

# ── Numerics / data ───────────────────────────────────────────────────────────
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
scipy>=1.12.0

# ── Video assembly ────────────────────────────────────────────────────────────
ffmpeg-python>=0.2.0         # Python wrapper for FFmpeg CLI
imageio[ffmpeg]>=2.34.0

# ── SoccerNet data toolkit ────────────────────────────────────────────────────
SoccerNet>=0.1.51

# ── Configuration & utilities ─────────────────────────────────────────────────
PyYAML>=6.0
tqdm>=4.66.0
rich>=13.7.0                 # Pretty console output

# ── Experiment tracking ───────────────────────────────────────────────────────
tensorboard>=2.16.0

# ── Notebook / analysis ───────────────────────────────────────────────────────
jupyter>=1.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
ipywidgets>=8.1.0
